ref: Move shutdown message to controller and organize utils
