+++ server/src/lib.rs
@@ -42,7 +42,6 @@ mod endpoints;
 mod extractors;
 mod middlewares;
 mod service;
-mod shutdown;
 mod utils;
 
 use actors::controller::Controller;
